move FanP9 and FanP10 to fan_miot.py
